meta: Moved quest stringify utilities into shared module
