+++ packages/gcdata/src/cl2.quest.stringify.ts
@@ -1,5 +1,5 @@
 import type { GameChanger } from './GameChanger.js';
-import { assert } from './assert.js';
+import { characterString, emojiString } from './cl2.shared.stringify.js';
 import type { Crashlands2 } from './cl2.types.auto.js';
 import { bsArrayToArray, toArrayTag, toMoteTag } from './helpers.js';
 import type { Mote } from './types.js';
@@ -106,9 +106,9 @@ export function stringifyQuest(
         if (moment.style === 'Dialogue') {
           // Speaker and dialog line
           if (moment.speech.speaker !== lastSpeaker) {
-            line += `\t${characterString(moment.speech.speaker)}\n`;
+            line += `\t${characterString(moment.speech.speaker, packed)}\n`;
           }
-          const emojiStr = emojiString(moment.speech.emotion);
+          const emojiStr = emojiString(moment.speech.emotion, packed);
           line += `>${toArrayTag(momentContainer)} ${
             emojiStr ? emojiStr + ' ' : ''
           }${moment.speech.text.text}`;
@@ -120,7 +120,8 @@ export function stringifyQuest(
             emojiLines.push(
               `!${toArrayTag(emote)} ${characterString(
                 emote.element?.key!,
-              )} ${emojiString(emote.element?.value)}`,
+                packed,
+              )} ${emojiString(emote.element?.value, packed)}`,
             );
           }
           line += emojiLines.join('\n');
@@ -172,21 +173,6 @@ export function stringifyQuest(
         }
       }
     }
-  }
-
-  function emojiString(emojiId: string | undefined) {
-    if (!emojiId) return '';
-    const emoji = packed.working.getMote(emojiId);
-    const name = packed.working.getMoteName(emoji) || emoji?.id || 'UNKNOWN';
-    return name ? `(${name})` : '';
-  }
-
-  function characterString(characterId: string) {
-    assert(characterId, 'Character ID must be defined');
-    const character = packed.working.getMote(characterId);
-    const name =
-      packed.working.getMoteName(character) || character?.id || 'UNKNOWN';
-    return name ? `${name.toUpperCase()}${toMoteTag(characterId)}` : '';
   }
 
   return blocks.join('\n') + '\n';
